backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
qdrant-client>=1.7.0
asyncpg>=0.29.0
openai>=1.3.5
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.23
alembic>=1.13.1
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
requests>=2.31.0
google-generativeai>=0.3.0